Move error code test to pkg/handlerutil
https://github.com/kubeapps/kubeapps/pull/1311#discussion_r349671189
