+++ cmd/tiller-proxy/internal/handler/handler_test.go
@@ -19,7 +19,6 @@ package handler
 import (
 	"context"
 	"fmt"
-	"net/http"
 	"net/http/httptest"
 	"reflect"
 	"strings"
@@ -31,31 +30,8 @@ import (
 	"github.com/kubeapps/kubeapps/pkg/auth"
 	authFake "github.com/kubeapps/kubeapps/pkg/auth/fake"
 	chartFake "github.com/kubeapps/kubeapps/pkg/chart/fake"
-	"github.com/kubeapps/kubeapps/pkg/handlerutil"
 	proxyFake "github.com/kubeapps/kubeapps/pkg/proxy/fake"
 )
-
-func TestErrorCodeWithDefault(t *testing.T) {
-	type test struct {
-		err          error
-		defaultCode  int
-		expectedCode int
-	}
-	tests := []test{
-		{fmt.Errorf("a release named foo already exists"), http.StatusInternalServerError, http.StatusConflict},
-		{fmt.Errorf("release foo not found"), http.StatusInternalServerError, http.StatusNotFound},
-		{fmt.Errorf("Unauthorized to get release foo"), http.StatusInternalServerError, http.StatusForbidden},
-		{fmt.Errorf("release \"Foo \" failed"), http.StatusInternalServerError, http.StatusUnprocessableEntity},
-		{fmt.Errorf("This is an unexpected error"), http.StatusInternalServerError, http.StatusInternalServerError},
-		{fmt.Errorf("This is an unexpected error"), http.StatusUnprocessableEntity, http.StatusUnprocessableEntity},
-	}
-	for _, s := range tests {
-		code := handlerutil.ErrorCodeWithDefault(s.err, s.defaultCode)
-		if code != s.expectedCode {
-			t.Errorf("Expected '%v' to return code %v got %v", s.err, s.expectedCode, code)
-		}
-	}
-}
 
 func TestActions(t *testing.T) {
 	type testScenario struct {
